requirements.txt
#weconnect-cupra-daern[Images]
paho-mqtt
